+++ R/editing.R
@@ -82,24 +82,46 @@ channels_tbl.data.frame <- function(x, ...) {
 `channels_tbl<-` <- function(x, value) {
   UseMethod("channels_tbl<-")
 }
+#' @rdname channels_tbl
+#' @export
+`channels_tbl<-.ica_lst` <- function(x, value) {
+  stop("Channels can't be edited in an ica_lst object.")
+  # orig_names <- channel_names(x$mixing)
+  # channels_mx <- x$mixing[, channel_names(x$mixing),, with =FALSE]
+  # nochannels_mx <- x$mixing[, setdiff(channel_names(x$mixing),colnames(x$mixing)), with = FALSE]
+  # x$mixing <- cbind(nochannels_mx, 
+  #                   data.table::as.data.table(
+  #                     update_channel_meta_data(channels_mx, 
+  #                                              value[value$.channel %in% colnames(x$mixing),]))) %>%
+  #   as_mixing_tbl()
+  # new_names <- channel_names(x$mixing)
+  # x$signal <- data.table::copy(x$signal)
+  # for (i in seq_len(nchannels(x$mixing))) {
+  #   data.table::set(x$signal,which(x$signal$.channel == orig_names[i]), ".channel", new_names[i])
+  # }
+  # NextMethod()
+  
+  }
+
 
 #' @rdname channels_tbl
 #' @export
 `channels_tbl<-.eeg_lst` <- function(x, value) {
-  orig_names <- channel_names(x)
-  channels <- dplyr::select(x$signal, orig_names)
-  nochannels <- dplyr::select(x$signal, -dplyr::one_of(channel_names(x)))
-  x$signal <- dplyr::bind_cols(nochannels, update_channel_meta_data(channels, value))
-  new_names <- channel_names(x)
-
-  for (i in seq_len(nchannels(x))) {
-    #TODO do it in data.table
-    x$events <- dplyr::mutate(x$events, 
-                              .channel = dplyr::if_else(.channel == orig_names[i], new_names[i], .channel)) %>%
-                as_events_tbl()
+  orig_names <- channel_names(x$signal)
+  channels_sg <- x$signal[, channel_names(x$signal), with =FALSE]
+  nochannels_sg <- x$signal[, setdiff(colnames(x$signal), channel_names(x$signal)), with = FALSE]
+  x$signal <- cbind(nochannels_sg, 
+                    data.table::as.data.table(
+                      update_channel_meta_data(channels_sg, 
+                                               value))) %>%
+    as_signal_tbl()
+  new_names <- channel_names(x$signal)
+  
+  x$signal <- data.table::copy(x$signal)
+  for (i in seq_len(nchannels(x$signal))) {
+    data.table::set(x$signal,which(x$signal$.channel == orig_names[i]), ".channel", new_names[i])
   }
-
-  x
+x
 }
 
 #' @rdname channels_tbl
@@ -110,4 +132,37 @@ channels_tbl.data.frame <- function(x, ...) {
   nochannels <- dplyr::select(x, -dplyr::one_of(channel_names(x)))
   x <- dplyr::bind_cols(nochannels, update_channel_meta_data(channels, value))
   x
+}
+#' 
+#' #' @export
+#' `[[.eeg_lst` <- function(x,i,...) {
+#'   pull.eeg_lst(.data = x,var =  i)
+#' }
+
+
+
+#' Function to get the signal table of an eeg_lst object.
+#'
+#'
+#' @param x An eeg_lst object.
+#' @param ... Not in use.
+#' 
+#' @return A table.
+#' @export
+signal_tbl <- function(x, ...) {
+    UseMethod("signal_tbl")
+}
+#' @rdname signal
+#' @export
+signal_tbl.eeg_lst <- function(x,...){
+    x$signal
+}
+#' @export
+`signal_tbl<-` <- function(x, value) {
+    UseMethod("signal_tbl<-")
+}
+#' @export
+`signal_tbl<-.eeg_lst` <- function(x, value) {
+
+stop("Not implemented, please use mutate", call. = FALSE)
 }