+++ R/read_helpers.R
@@ -132,7 +132,7 @@ read_dat <- function(file, header_info = NULL, events = NULL,
 
 
 add_event_channel <- function(events, labels) {
-  labels <- make.names(labels)
+  labels <- make_names(labels)
   dplyr::mutate(events, .channel = if (".channel" %in% names(events)) {
     .channel
   } else {
@@ -195,7 +195,16 @@ read_vmrk <- function(file) {
   # splits Mk<Marker number>=<Type>, removes the Mk.., and <Date>
   events[,type := stringr::str_split(type,"=") %>%
            purrr::map_chr(~.x[[2]])][,date := NULL]
-  events[, .final := .initial + .final - 1L][]
+    events[, .final := .initial + .final -1L]
+
+
+    ## bug of BV1, first sample is supposed to be 1
+    BV1 <- "BrainVision Data Exchange Marker File, Version 1.0"
+    if(stringr::str_detect(readChar(file, nchar(BV1)),BV1) &&
+       events[type=="New Segment", .initial][1]==0){
+        events[type=="New Segment" & .initial==0, c(".initial",".final") := list(1,1)]
+    }
+    events
 }
 
 #' @importFrom magrittr %>%