+++ R/to_tbl.R
@@ -9,7 +9,7 @@
 #' 
 #'
 #'
-as.data.table.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_channels_info = TRUE) {
+as.data.table.eeg_lst <- function(x, unit = "second") {
     ## check if there are channels in the signal tbl (in ica_lst they may not be there)
     if(any(channel_names(x) %in% colnames(x$signal))){
         ## channels <- x$signal %>% dplyr::select_at(vars(-one_of(component_names(x)))) %>%
@@ -17,13 +17,15 @@ as.data.table.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_c
         ##     tidyr::gather(key = ".source", value = ".value",
         ##                   (intersect(colnames(x$signal), channel_names(x)))) %>%
         ##     dplyr::mutate(.type = "channel")
-
-        channels <- x$signal %>% dplyr::select_at(vars(-one_of(component_names(x)))) %>%
+  
+        channels <- x$signal %>% dplyr::select_at(dplyr::vars(-dplyr::one_of(component_names(x)))) %>%
             data.table::melt(variable.name = ".source",
                              measure.vars = intersect(colnames(x$signal), channel_names(x)),
                              value.name = ".value")
         channels[,.type := "channel"][
-           ,.source := as.character(.source)]
+           ,.source := as.character(.source)][
+           ,.value := `attributes<-`(.value,NULL)   
+           ]
     } else {
         channels <- data.table::data.table()
     }
@@ -33,32 +35,20 @@ as.data.table.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_c
         ##     .[,lapply(.SD, `attributes<-`, NULL )] %>%
         ##     tidyr::gather(key = ".source", value = ".value", component_names(x)) %>%
         ##     dplyr::mutate(.type = "component")
-        components <- x$signal %>% dplyr::select_at(vars(one_of(component_names(x)))) %>%
+        components <- x$signal %>% dplyr::select_at(dplyr::vars(-dplyr::one_of(channel_names(x$signal)))) %>%
             data.table::melt(variable.name = ".source",
                              measure.vars = component_names(x),
                              value.name = ".value")
         components[,.type := "component"][
-           ,.source := as.character(.source)]
+           ,.source := as.character(.source)][
+           ,.value := `attributes<-`(.value,NULL)   
+           ]
     } else {
         components =  data.table::data.table()
     }
 
     long_table <- rbind(channels,components) %>%
-        ## remove attributes that are now problematic
-        {
-            if (add_segments) {
                 left_join_dt(., data.table::as.data.table(x$segments), by = ".id")
-            } else {
-                .
-            }
-        } %>% 
-        {
-            if (add_channels_info) {
-                left_join_dt(., data.table::as.data.table(channels_tbl(x)), by = c(".source"=".channel"))
-            } else {
-                .
-            }
-        }
 
     ## inf_events <- setdiff(colnames(events_tbl(x)),  obligatory_cols[["events"]])
     ## events_all_ch <- data.table::copy(events_tbl(x))[is.na(.channel), ends := .initial + .size]
@@ -69,8 +59,9 @@ as.data.table.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_c
     .unit <- unit
     long_table[, time := as_time(.sample_id, unit = .unit)]
     long_table[, .sample_id := NULL]
-    long_table %>% dplyr::select(time, dplyr::everything()) %>%
-    .[,lapply(.SD, `attributes<-`, NULL )]
+    long_table %>% dplyr::select(time, dplyr::everything())
+    #%>%
+    #.[,lapply(.SD, `attributes<-`, NULL )]
 }
 
 
@@ -86,8 +77,8 @@ as.data.table.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_c
 #'
 #' @family tibble
 #'
-as_tibble.eeg_lst <- function(x, unit = "second", add_segments = TRUE, add_channels_info = TRUE) {
-    data.table::as.data.table(x, unit, add_segments, add_channels_info) %>%
+as_tibble.eeg_lst <- function(x, unit = "second") {
+    data.table::as.data.table(x, unit) %>%
         dplyr::as_tibble()
 }