+++ data-raw/create_layout.R
@@ -1,0 +1,5 @@
+library(eeguana)
+
+layout_32_1020 <- readr::read_csv("data-raw/layout_32_1020.csv")
+
+usethis::use_data(layout_32_1020, overwrite=TRUE)